Remove PhoneUtils, Country Functions and io from celo/utils
