+++ packages/attestation-service/package.json
@@ -32,6 +32,7 @@
     "@celo/contractkit": "1.3.1-dev",
     "@celo/keystores": "1.3.1-dev",
     "@celo/utils": "1.3.1-dev",
+    "@celo/phone-utils": "1.3.1-dev",
     "bignumber.js": "^9.0.0",
     "body-parser": "1.19.0",
     "bunyan": "1.8.12",
