+++ packages/attestation-service/src/requestHandlers/attestation.ts
@@ -1,8 +1,9 @@
 import { AttestationState } from '@celo/contractkit/lib/wrappers/Attestations'
-import { AttestationUtils, PhoneNumberUtils } from '@celo/utils'
+import { PhoneNumberUtils } from '@celo/phone-utils'
+import { AttestationRequest } from '@celo/phone-utils/lib/io'
+import { AttestationUtils } from '@celo/utils'
 import { eqAddress } from '@celo/utils/lib/address'
 import { sleep } from '@celo/utils/lib/async'
-import { AttestationRequest } from '@celo/utils/lib/io'
 import Logger from 'bunyan'
 import { randomBytes } from 'crypto'
 import express from 'express'
