+++ packages/env-tests/src/shared/attestation.ts
@@ -8,9 +8,9 @@ import {
 } from '@celo/contractkit/lib/wrappers/Attestations'
 import { OdisUtils } from '@celo/identity'
 import { AuthSigner } from '@celo/identity/lib/odis/query'
-import { AttestationUtils, PhoneNumberUtils } from '@celo/utils'
+import { AttestationRequest, PhoneNumberUtils } from '@celo/phone-utils'
+import { AttestationUtils } from '@celo/utils'
 import { concurrentMap, sleep } from '@celo/utils/lib/async'
-import { AttestationRequest } from '@celo/utils/lib/io'
 import Logger from 'bunyan'
 import { sample } from 'lodash'
 import moment from 'moment'
