+++ packages/phone-number-privacy/combiner/test/end-to-end/resources.ts
@@ -5,7 +5,7 @@ import {
   ServiceContext,
   WalletKeySigner,
 } from '@celo/identity/lib/odis/query'
-import { PhoneNumberUtils } from '@celo/utils'
+import { PhoneNumberUtils } from '@celo/phone-utils'
 import {
   ensureLeading0x,
   normalizeAddressWith0x,
