+++ packages/sdk/contractkit/src/wrappers/Attestations.ts
@@ -4,8 +4,8 @@ import { notEmpty, zip3 } from '@celo/base/lib/collections'
 import { parseSolidityStringArray } from '@celo/base/lib/parsing'
 import { appendPath } from '@celo/base/lib/string'
 import { Address, toTransactionObject } from '@celo/connect'
+import { AttestationRequest, GetAttestationRequest } from '@celo/phone-utils/lib/io'
 import { AttestationUtils, SignatureUtils } from '@celo/utils/lib'
-import { AttestationRequest, GetAttestationRequest } from '@celo/utils/lib/io'
 import { attestationSecurityCode as buildSecurityCodeTypedData } from '@celo/utils/lib/typed-data-constructors'
 import BigNumber from 'bignumber.js'
 import fetch from 'cross-fetch'
