+++ packages/sdk/identity/package.json
@@ -27,6 +27,7 @@
   "dependencies": {
     "@celo/base": "1.3.1-dev",
     "@celo/utils": "1.3.1-dev",
+    "@celo/phone-utils": "1.3.1-dev",
     "@celo/contractkit": "1.3.1-dev",
     "@types/debug": "^4.1.5",
     "bignumber.js": "^9.0.0",
