+++ packages/sdk/identity/src/odis/matchmaking.ts
@@ -1,4 +1,4 @@
-import { E164Number } from '@celo/utils/lib/io'
+import { E164Number } from '@celo/phone-utils/lib/io'
 import crypto from 'crypto'
 import debugFactory from 'debug'
 import {
