+++ packages/sdk/phone-utils/package.json
@@ -25,18 +25,12 @@
     "@types/google-libphonenumber": "^7.4.17",
     "@types/node": "^10.12.18",
     "country-data": "^0.0.31",
-    "ethereumjs-util": "^5.2.0",
     "fp-ts": "2.1.1",
     "io-ts": "2.0.1",
-    "google-libphonenumber": "^3.2.15",
-    "web3-utils": "1.3.6"
+    "google-libphonenumber": "^3.2.15"
   },
   "devDependencies": {
     "@celo/flake-tracker": "0.0.1-dev",
     "@celo/typescript": "0.0.1"
-  },
-  "resolutions": {
-    "elliptic/bn.js": "4.11.8",
-    "web3-utils/bn.js": "4.11.8"
   }
 }
