+++ packages/sdk/phone-utils/src/index.ts
@@ -1,4 +1,19 @@
 export * from './countries'
+export { Countries, LocalizedCountry } from './countries'
+export * from './getCountryEmoji'
+export * from './getPhoneHash'
 export * from './inputValidation'
 export * from './io'
-export { PhoneNumberUtils } from './phoneNumbers'
+export {
+  getCountryCode,
+  getDisplayNumberInternational,
+  getDisplayPhoneNumber,
+  getE164DisplayNumber,
+  getE164Number,
+  getExampleNumber,
+  getRegionCode,
+  getRegionCodeFromCountryCode,
+  isE164NumberStrict,
+  parsePhoneNumber,
+  PhoneNumberUtils,
+} from './phoneNumbers'
