+++ packages/sdk/phone-utils/src/io.ts
@@ -1,40 +1,7 @@
-import { URL_REGEX } from '@celo/base/lib/io'
-import { isValidAddress } from '@celo/utils/lib/address'
-import { isValidPublic, toChecksumAddress } from 'ethereumjs-util'
+import { AddressType, SaltType, SignatureType } from '@celo/utils/lib/io'
 import { either } from 'fp-ts/lib/Either'
 import * as t from 'io-ts'
 import { isE164NumberStrict } from './phoneNumbers'
-
-// Exports moved to @celo/base, forwarding them
-// here for backwards compatibility
-export { isValidUrl, URL_REGEX } from '@celo/base/lib/io'
-
-export const UrlType = new t.Type<string, string, unknown>(
-  'Url',
-  t.string.is,
-  (input, context) =>
-    either.chain(t.string.validate(input, context), (stringValue) =>
-      URL_REGEX.test(stringValue)
-        ? t.success(stringValue)
-        : t.failure(stringValue, context, 'is not a valid url')
-    ),
-  String
-)
-
-export const JSONStringType = new t.Type<string, string, unknown>(
-  'JSONString',
-  t.string.is,
-  (input, context) =>
-    either.chain(t.string.validate(input, context), (stringValue) => {
-      try {
-        JSON.parse(stringValue)
-        return t.success(stringValue)
-      } catch (error) {
-        return t.failure(stringValue, context, 'can not be parsed as JSON')
-      }
-    }),
-  String
-)
 
 export const E164PhoneNumberType = new t.Type<string, string, unknown>(
   'E164Number',
@@ -47,35 +14,6 @@ export const E164PhoneNumberType = new t.Type<string, string, unknown>(
     ),
   String
 )
-
-export const AddressType = new t.Type<string, string, unknown>(
-  'Address',
-  t.string.is,
-  (input, context) =>
-    either.chain(t.string.validate(input, context), (stringValue) =>
-      isValidAddress(stringValue)
-        ? t.success(toChecksumAddress(stringValue))
-        : t.failure(stringValue, context, 'is not a valid address')
-    ),
-  String
-)
-
-export const PublicKeyType = new t.Type<string, string, unknown>(
-  'Public Key',
-  t.string.is,
-  (input, context) =>
-    either.chain(t.string.validate(input, context), (stringValue) =>
-      stringValue.startsWith('0x') && isValidPublic(Buffer.from(stringValue.slice(2), 'hex'), true)
-        ? t.success(toChecksumAddress(stringValue))
-        : t.failure(stringValue, context, 'is not a valid public key')
-    ),
-  String
-)
-
-export const SignatureType = t.string
-
-export const SaltType = t.string
-
 export const AttestationServiceStatusResponseType = t.type({
   status: t.literal('ok'),
   smsProviders: t.array(t.string),
@@ -101,8 +39,6 @@ export const AttestationServiceTestRequestType = t.type({
 })
 export type AttestationServiceTestRequest = t.TypeOf<typeof AttestationServiceTestRequestType>
 
-export type Signature = t.TypeOf<typeof SignatureType>
-export type Address = t.TypeOf<typeof AddressType>
 export type E164Number = t.TypeOf<typeof E164PhoneNumberType>
 
 export const AttestationRequestType = t.type({
