+++ packages/sdk/phone-utils/src/phoneNumbers.ts
@@ -1,49 +1,14 @@
-import {
-  getPhoneHash as baseGetPhoneHash,
-  isE164Number,
-  ParsedPhoneNumber,
-} from '@celo/base/lib/phoneNumbers'
-import CountryData from 'country-data'
+import { isE164Number, ParsedPhoneNumber } from '@celo/base/lib/phoneNumbers'
 import {
   PhoneNumber,
   PhoneNumberFormat,
   PhoneNumberType,
   PhoneNumberUtil,
 } from 'google-libphonenumber'
-import { soliditySha3 } from 'web3-utils'
-
-// Exports moved to @celo/base, forwarding them
-// here for backwards compatibility
-export { anonymizedPhone, isE164Number, ParsedPhoneNumber } from '@celo/base/lib/phoneNumbers'
-
-const sha3 = (v: string): string | null => soliditySha3({ type: 'string', value: v })
-export const getPhoneHash = (phoneNumber: string, salt?: string): string => {
-  return baseGetPhoneHash(sha3, phoneNumber, salt)
-}
+import getPhoneHash from './getPhoneHash'
 
 const phoneUtil = PhoneNumberUtil.getInstance()
 const MIN_PHONE_LENGTH = 4
-
-export function getCountryEmoji(
-  e164PhoneNumber: string,
-  countryCodePossible?: number,
-  regionCodePossible?: string
-) {
-  // The country code and region code can both be passed in, or it can be inferred from the e164PhoneNumber
-  let countryCode: any
-  let regionCode: any
-  countryCode = countryCodePossible
-  regionCode = regionCodePossible
-  if (!countryCode || !regionCode) {
-    countryCode = getCountryCode(e164PhoneNumber)
-    regionCode = getRegionCode(e164PhoneNumber)
-  }
-  const countries = CountryData.lookup.countries({ countryCallingCodes: `+${countryCode}` })
-  const userCountryArray = countries.filter((c: any) => c.alpha2 === regionCode)
-  const country = userCountryArray.length > 0 ? userCountryArray[0] : undefined
-
-  return country ? country.emoji : ''
-}
 
 export function getCountryCode(e164PhoneNumber: string) {
   if (!e164PhoneNumber) {
