+++ packages/sdk/phone-utils/tsconfig.json
@@ -6,5 +6,6 @@
     "downlevelIteration": true,
     "resolveJsonModule": true
   },
-  "include": ["src/", "src/**/*.json"]
+  "include": ["src/**/*.ts", "src/**/*.json"],
+  "exclude": ["src/**/*.test.ts"]
 }
