+++ packages/sdk/utils/package.json
@@ -20,10 +20,8 @@
   ],
   "dependencies": {
     "@celo/base": "1.3.1-dev",
-    "@types/country-data": "^0.0.0",
     "@types/elliptic": "^6.4.9",
     "@types/ethereumjs-util": "^5.2.0",
-    "@types/google-libphonenumber": "^7.4.17",
     "@types/node": "^10.12.18",
     "@types/randombytes": "^2.0.0",
     "bigi": "^1.1.0",
@@ -33,12 +31,9 @@
     "bls12377js": "https://github.com/celo-org/bls12377js#cb38a4cfb643c778619d79b20ca3e5283a2122a6",
     "bn.js": "4.11.8",
     "buffer-reverse": "^1.0.1",
-    "country-data": "^0.0.31",
     "crypto-js": "^3.1.9-1",
     "elliptic": "^6.5.4",
     "ethereumjs-util": "^5.2.0",
-    "fp-ts": "2.1.1",
-    "google-libphonenumber": "^3.2.15",
     "io-ts": "2.0.1",
     "keccak256": "^1.0.0",
     "numeral": "^2.0.6",
