+++ packages/sdk/utils/src/index.ts
@@ -4,13 +4,11 @@ export { AccountUtils } from './account'
 export { AttestationUtils } from './attestations'
 export { CommentEncryptionUtils } from './commentEncryption'
 export * from './contacts'
-export * from './countries'
 export * from './currencies'
 export * from './dappkit'
 export { DataEncryptionKeyUtils } from './dataEncryptionKey'
 export { ECIES } from './ecies'
 export { IstanbulUtils } from './istanbul'
-export { PhoneNumberUtils } from './phoneNumbers'
 export { SignatureUtils } from './signatureUtils'
 export { StringUtils } from './string'
 export { AddressUtils }
